Style
scss

+++ src/controllers/videoController.js
@@ -1,9 +1,11 @@
 import Video from "../models/Video";
+import User from "../models/User";
+import { isValidObjectId } from "mongoose";
 
 
 export const home = async(req,res) => {
     try{
-        const videos = await Video.find({}).sort({ createdAt: "ascending"});
+        const videos = await Video.find({}).sort({ createdAt: "ascending"}).populate("owner");
         return res.render("home",{pageTitle:"Home", videos });
     }
     catch {
@@ -12,7 +14,7 @@ export const home = async(req,res) => {
 };
 export const watch = async (req,res) => {
     const {id} = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video) {
         return res.render("404", {pageTitle:"Video not found."});
     } 
@@ -20,19 +22,27 @@ export const watch = async (req,res) => {
 }
 export const getEdit = async(req,res) => {
     const {id} = req.params;
+    const {user :{_id}} = req.session
     const video = await Video.findById(id);
     if(!video) {
         return res.status(404).render("404", {pageTitle:"Video not found."});
     } 
+    if(Sting(video.owner) !== Stirng(_id)) {
+        return res.status(403).redirect("/");
+    }
     return res.render("edit", {pageTitle: `Edit ${video.title}`,video});
 }
 export const postEdit = async (req,res) => {
     const {id} = req.params;
+    const {user :{_id}} = req.session
     const{title, description, hashtags} = req.body
-    const video = await Video.exists({_id:id});
+    const video = await Video.findById(id);
     if(!video) {
         return res.status(404).render("404", {pageTitle:"Video not found."});
     } 
+    if(Sting(video.owner) !== Stirng(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndUpdate(id,{
         title,
         description,
@@ -46,13 +56,20 @@ export const getUpload = (req,res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const {user: {_id}} = req.session;
+    const {path: fileUrl} = req.file
     const {title, description, hashtags} = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl ,
+            owner: _id ,
             hashtags: Video.formatHashtags(hashtags)
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id)
+        user.save();
         return res.redirect("/");
     } catch(error) {
         return res.status(400).render("upload",{pageTitle: "Upload Video", errorMessage: error._message,});
@@ -61,9 +78,22 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req,res) => {
     const {id} = req.params;
+    const {user :{_id}} = req.session
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+
+    if(!video) {
+        return res.status(404).render("404", {pageTitle:"Video not found."});
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    user.videos.splice(user.videos.indexOf(id),1);
+    user.save();
     return res.redirect("/");
 };
+
 
 export const search = async (req,res) => {
     const {keyword} = req.query;
@@ -73,7 +103,7 @@ export const search = async (req,res) => {
             title: {
                 $regex: new RegExp(keyword,"i")
             }
-        });
+        }).populate("owner");
     }
     return res.render("search",{pageTitle:"Search",videos});
 }
